#67 - splitting dashboard file into components
